+++ polaris-domain/src/main/java/com/achobeta/domain/team/model/entity/PositionEntity.java
@@ -30,7 +30,7 @@ public class PositionEntity {
     /**
      * 团队架构中的等级 0-根节点/团队 1-一级节点 2-二级节点 3-三级节点 4-四级节点
      */
-    private Byte level;
+    private Integer level;
     /**
      * 层级关系的子节点/下级分组
      */
